Bootstrap a DEA model

Run bootstrap on a DEA model to estimate bias corrected efficiency scores and confidence intervals.

Usage

bootstrap_dea(dea, alpha = 0.05, bw_rule = "ucv", iterations = 2000)

Arguments

dea

An object of class 'pioneer_dea' from compute_dea().

alpha

One minus the confidence level required. Default is 0.05.

bw_rule

A string with the type of bandwidth rule to be used, or a number with the bandwidth parameter. See details.

iterations

The number of bootstrap iterations to perform. Default is 2000.

Value

A list of class pioneer_bootstrap.

Details

In order to bootstrap a DEA model, you must first create a model object using the compute_dea() function. Note that you currently can only bootstrap models using constant or variable returns to scale (RTS). If you try to bootstrap a model using another RTS, the bootstrap will fail with an error message.

The bandwidth argument can be set to either ucv for unbiased cross validation, silverman for the Silverman rule, or scott for the Scott rule. If you provide a number, this will be used directly as the bandwidth parameter h. This can be useful to replicate results where h is given, such as Simar & Wilson (1998). For most practical applications of the bootstrap, the default value of unbiased cross validation is sensible.

See also

Examples

# Load example data
fare89 <- deaR::Electric_plants
# Estimate efficiency
mod <- compute_dea(
  data = fare89,
  input = c("Labor", "Fuel", "Capital"),
  output = "Output",
  id = "Plant",
)
# Run bootstrap. Reducing the number of iterations to save processing time
boot <- bootstrap_dea(mod, iterations = 100)
# Print results
print(boot)
#> Bias corrected efficiency scores:
#>  [1] 0.8269453 0.8736980 0.8127673 0.8719568 0.8890060 0.8375015 0.8500982
#>  [8] 0.7802279 0.8216964 0.8063358 0.9197955 0.8638909 0.9169919 0.8069850
#> [15] 0.8723040 0.8281121 0.9070702 0.7872815 0.9089161
# Get summary
summary(boot)
#> Bootstrap with 100 iterations of DEA model with technology CRS and input oriented efficieny
#> Mean bias corrected efficiency: 0.8517
#> Mean bias: 0.045
#> -----------
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>  0.7802  0.8172  0.8501  0.8517  0.8814  0.9198