get_auth_token {pioneer}	R Documentation

get_auth_token function to authenticate with the Pioneer TRM Tracker API.

Description

The function, using your Pioneer login credentials, retrieves the required Auth token to pull data/reports.

Use this function to assign the Auth token to an object.

While not necessary, you may want to store your API credentials in your .Renviron file. If so, place "pioneer_usr='<YOUR_USERNAME>'" & "pioneer_pwd='<YOUR_PASSWORD>'" into .Renviron

Usage

get_auth_token(username = Sys.getenv("pioneer_usr"), password = Sys.getenv("pioneer_pwd"), verbose = TRUE)

Arguments

username
Your username for Pioneer's API - the same one you login with to the TRMTracker website The function is built to automatically pull your username from your .Renviron file.

password
Your password for Pioneer's API, the same one you login with to the TRMTracker website. The function is built to automatically pull your password from your .Renviron file.

verbose
Retrieves, and prints to console, all HTTP traffic information. This has been directly inherited from the 'httr' package.

Examples

## The function is currently defined as
myToken <- get_auth_token(username = 'XXXX', password = 'XXXXXXX', verbose = T)

[Package pioneer version 0.1.0 Index]