get_import_format {pioneer}	R Documentation

get_import_format

Description

This function returns, as a data.frame, the list of available Pioneer reports and their Hash values

Usage

get_import_format(tokenID, verbose=TRUE)

Arguments

tokenID
This is your Auth token, a result from the 'get_auth_token()' function

verbose
Retrieves, and prints to console, all HTTP traffic information. This has been directly inherited from the 'httr' package.

Value

An n*1 data.frame with each column providing the title for that column for input

Examples

## The function is currently defined as
myImportFormat <- get_import_format(tokenID = myToken, verbose = T)

[Package pioneer version 0.1.0 Index]