get_report_data {pioneer}	R Documentation

get_report_data

Usage

get_report_data(tokenID, report_parameters, verbose = TRUE)

Arguments

tokenID
This is your Auth token, a result from the 'get_auth_token()' function

report_parameters
This is a list containing 3 objects:

report_name

report_hash

report_parameters

where 'report_parameters' is a data.frame consisting of the required parameters and their values for the query.

verbose
Retrieves, and prints to console, all HTTP traffic information. This has been directly inherited from the 'httr' package.

Value

A data.frame containing any and all data from the specified report given parameter limits.

Examples

# Authenticate
myToken <- get_auth_token()

# Pull list of reports
myReportList <- get_report_list(tokenID = myToken)

# Get the Price Report info (name and hash value)
myReport <- myReportList[which(myReportList$report_name == 'Price Report'),]

# Pull the required parameters for the specified report (Price Report) to later retrieve its data
reportParameters <- get_report_params(tokenID = myToken, report = myReport, verbose = T)

# Build report_parameters object
rpt_params <- list('report_name' = myReport$report_name,
                   'report_hash' = myReport$report_hash,
                   'report_parameters' = as.data.frame('channel' = 'NULL',
                                                      'hub' = 'NULL',
                                                      'uid' = 'NULL',
                                                      'zone' = 'NULL',
                                                      'date' = '04/30/2020',
                                                      stringsAsFactors = FALSE))

# Query the API
data <- get_report_data(tokenID = myToken, report_parameters = rpt_params, verbose = T)

[Package pioneer version 0.1.0 Index]