get_report_params {pioneer}	R Documentation

get_report_params

Usage

get_report_params(tokenID, report, verbose = TRUE)

Arguments

tokenID
This is your Auth token, a result from the 'get_auth_token()' function

report
A 1x2 data.frame from the report list, where the first column is the report name, and the second column is the report hash value

verbose
Retrieves, and prints to console, all HTTP traffic information. This has been directly inherited from the 'httr' package.

Value

An nx3 data.frame consisting of the required parameters, and their data types, for the specified report.

Examples

# Authenticate
myToken <- get_auth_token()

# Pull list of reports
myReportList <- get_report_list(tokenID = myToken)

# Get the Price Report info (name and hash value)
myReport <- myReportList[which(myReportList$report_name == 'Price Report'),]

# Pull the required parameters for the specified report (Price Report) to later retrieve its data
reportParameters <- get_report_params(tokenID = myToken, report = myReport, verbose = T)

[Package pioneer version 0.1.0 Index]