import_data {pioneer}	R Documentation

import_data

Description

This function returns, as a data.frame, the list of available Pioneer reports and their Hash values

Usage

import_data(tokenID, importRule, data, verbose=TRUE)

Arguments

tokenID
This is your Auth token, a result from the 'get_auth_token()' function.

importRule
This is the name of the import rule/function for which you are choosing to import data thru.

data
This is a data.frame containing the data you wish you import in the format required to import. It will be converted into a JSON format for importing.

verbose
Retrieves, and prints to console, all HTTP traffic information. This has been directly inherited from the 'httr' package.

Examples

## The function is currently defined as
import_data(tokenID = myToken, importRule = 'PriceCurve', data = PriceCurveData, verbose = T)

[Package pioneer version 0.1.0 Index]