Create a summary table for DEA

Create a binned summary table for efficiency scores from a DEA model.

Usage

summary_tbl_dea(x)

Arguments

x

A vector of efficiency scores or a Farrell object

Value

data.frame

Details

The function will return a summary table for efficiency scores from a DEA model. Efficiency scores will be placed in 11 bins, where DMUs with an effciency score equal to 1 are placed in a separate bin. For output oriented models with range [1, Inf], bins are created with 1/bin. Bin widths will be equal to models with range [0, 1].