Create a summary table for DEA

Create a binned summary table for efficiency scores from a DEA model.

Usage

summary_tbl_dea(x)

Arguments

x

A numeric vector of efficiency scores or an object of class pioneer_dea

Value

A data.frame() with summary statistics

Details

The function will return a summary table for efficiency scores from a DEA model. Efficiency scores will be placed in 11 bins, where DMUs with an efficiency score equal to 1 are placed in a separate bin. For output oriented models with range [1, Inf], bins are created with 1/bin. Bin widths will be equal to models with range [0, 1].

Examples

# Load example data
fare89 <- deaR::Electric_plants
# Estimate efficiency
mod <- compute_dea(
  data = fare89,
  input = c("Labor", "Fuel", "Capital"),
  output = "Output",
  rts = "vrs"
)
# Get a summary table of efficiency scores
summary_tbl_dea(mod)
#>             Range Frequency
#> 1    0 <= E < 0.1         0
#> 2  0.1 <= E < 0.2         0
#> 3  0.2 <= E < 0.3         0
#> 4  0.3 <= E < 0.4         0
#> 5  0.4 <= E < 0.5         0
#> 6  0.5 <= E < 0.6         0
#> 7  0.6 <= E < 0.7         0
#> 8  0.7 <= E < 0.8         0
#> 9  0.8 <= E < 0.9         6
#> 10   0.9 <= E < 1         5
#> 11         E == 1         8
# You can also create the table from a numeric vector of efficiency scores
res <- as.data.frame(mod)
summary_tbl_dea(res$efficiency)
#>             Range Frequency
#> 1    0 <= E < 0.1         0
#> 2  0.1 <= E < 0.2         0
#> 3  0.2 <= E < 0.3         0
#> 4  0.3 <= E < 0.4         0
#> 5  0.4 <= E < 0.5         0
#> 6  0.5 <= E < 0.6         0
#> 7  0.6 <= E < 0.7         0
#> 8  0.7 <= E < 0.8         0
#> 9  0.8 <= E < 0.9         6
#> 10   0.9 <= E < 1         5
#> 11         E == 1         8